Reorganizing ActiveFedora:File code
This splits out some of the code in ActiveFedora::File into relevant modules for better organization and to
make the File class overall lighter and more easy to read.
